refactor llvm lowlevel code into its own file
