+++ crates/compiler/gen_llvm/src/llvm/mod.rs
@@ -6,4 +6,6 @@ pub mod compare;
 pub mod convert;
 mod expect;
 pub mod externs;
+mod intrinsics;
+mod lowlevel;
 pub mod refcounting;
